Make sure all logging happens in the composables
